// src/app/core/models/registration-model.ts
export class RegistrationModel{
        first_name: string;
        last_name: string;
        practice_name: string;
        address: string;
        state: string;
        city: string;
        country: string;
        zip: string;
        email_address: string;
        phone_number: string;
        user_name: string;
        password: string;
    }